Updated compilation configuration for x64 architecture, fixed 2 problems on x64 (dialog function declaration and disabled exceptions for float calculations)

+++ ResidFp/libsidplayfp/src/sidplayfp/sidbuilder.h
@@ -82,7 +82,7 @@ public:
      *
      * @return number of used sids, 0 if none.
      */
-    unsigned int usedDevices() const { return sidobjs.size(); }
+    size_t usedDevices() const { return sidobjs.size(); }
 
     /**
      * Available devices.
